Register "TransportOptions->Pinned Play/Record &Head (on/off)"; ...
... break TransportMenus -> AdornedRulerPanel
