+++ src/AdornedRulerPanel.cpp
@@ -2806,3 +2806,29 @@ void AdornedRulerPanel::TogglePinnedHead()
    if (scrubber.HasMark())
       scrubber.SetScrollScrubbing(value);
 }
+
+// Attach menu item
+
+#include "commands/CommandContext.h"
+#include "commands/CommandManager.h"
+#include "CommonCommandFlags.h"
+
+namespace {
+void OnTogglePinnedHead(const CommandContext &context)
+{
+   AdornedRulerPanel::Get( context.project ).TogglePinnedHead();
+}
+
+using namespace MenuTable;
+using Options = CommandManager::Options;
+AttachedItem sAttachment{
+   { wxT("Transport/Other/Options/Part2"), { OrderingHint::Begin, {} } },
+   Command( wxT("PinnedHead"), XXO("Pinned Play/Record &Head (on/off)"),
+      OnTogglePinnedHead,
+      // Switching of scrolling on and off is permitted
+      // even during transport
+      AlwaysEnabledFlag,
+      Options{}.CheckTest([](const AudacityProject&){
+         return TracksPrefs::GetPinnedHeadPreference(); } ) )
+};
+}
