+++ src/menus/TransportMenus.cpp
@@ -1,6 +1,3 @@
-
-
-#include "../AdornedRulerPanel.h"
 #include "../AudioIO.h"
 #include "../CommonCommandFlags.h"
 #include "DeviceManager.h"
@@ -410,11 +407,6 @@ void OnToggleSoundActivated(const CommandContext &WXUNUSED(context) )
    MenuManager::ModifyAllProjectToolbarMenus();
 }
 
-void OnTogglePinnedHead(const CommandContext &context)
-{
-   AdornedRulerPanel::Get( context.project ).TogglePinnedHead();
-}
-
 void OnTogglePlayRecording(const CommandContext &WXUNUSED(context) )
 {
    bool Duplex;
@@ -811,7 +803,7 @@ BaseItemSharedPtr TransportMenu()
             OnRescanDevices, AudioIONotBusyFlag() | CanStopAudioStreamFlag() ),
 
          Menu( wxT("Options"), XXO("Transport &Options"),
-            Section( "",
+            Section( "Part1",
                // Sound Activated recording options
                Command( wxT("SoundActivationLevel"),
                   XXO("Sound Activation Le&vel..."), OnSoundActivated,
@@ -823,15 +815,7 @@ BaseItemSharedPtr TransportMenu()
                   Options{}.CheckTest(wxT("/AudioIO/SoundActivatedRecord"), false) )
             ),
 
-            Section( "",
-               Command( wxT("PinnedHead"), XXO("Pinned Play/Record &Head (on/off)"),
-                  OnTogglePinnedHead,
-                  // Switching of scrolling on and off is permitted
-                  // even during transport
-                  AlwaysEnabledFlag,
-                  Options{}.CheckTest([](const AudacityProject&){
-                     return TracksPrefs::GetPinnedHeadPreference(); } ) ),
-
+            Section( "Part2",
                Command( wxT("Overdub"), XXO("&Overdub (on/off)"),
                   OnTogglePlayRecording,
                   AudioIONotBusyFlag() | CanStopAudioStreamFlag(),
